Hg issue 273: Missing unicode normalization quick check properties
The Unicode normalization quick check properties weren't handled correctly.

+++ regex_2/_regex.c
@@ -25439,7 +25439,7 @@ Py_LOCAL_INLINE(BOOL) init_property_dict(void) {
                 goto error;
         }
 
-        v = Py_BuildValue("i", value->id);
+        v = Py_BuildValue("i", (int)value->id);
         if (!v)
             goto error;
 
@@ -25461,7 +25461,7 @@ Py_LOCAL_INLINE(BOOL) init_property_dict(void) {
         int status;
 
         property = &re_properties[i];
-        v = Py_BuildValue("iO", property->id,
+        v = Py_BuildValue("iO", (int)property->id,
           value_dicts[property->value_set]);
         if (!v)
             goto error;
